+++ classes/inheritance/vehicle-five/vehicle_five.py
@@ -1,0 +1,5 @@
+from vehicle import Vehicle
+
+class Ferrari(Vehicle):
+    def __init__(self):
+        self.fuel_capacity = 0
