+++ classes/inheritance/vehicle-four/vehicle_four.py
@@ -1,0 +1,5 @@
+from vehicle import Vehicle
+
+class Porsche(Vehicle):
+    def __init__(self):
+        self.fuel_capacity = 0
