+++ classes/inheritance/vehicle-three/vehicle_three.py
@@ -1,0 +1,5 @@
+from vehicle import Vehicle
+
+class Ford(Vehicle):
+    def __init__(self):
+        self.fuel_capacity = 0
