+++ classes/inheritance/vehicle-two/vehicle_two.py
@@ -1,0 +1,5 @@
+from vehicle import Vehicle
+
+class Tesla(Vehicle):
+    def __init__(self):
+        self.battery_kwh = 0
